FrameChangeObserver - params and examples
Issue #9
Cleanup of class FrameChangeObserver.

Revamp class example so it will not choke on the PC.
It causes a MethodError because nil (from_scene) does not respond to
:name.
Add in a if conditional to only print the message once, ...
then thereafter draw a progress bar to the status line, ...
and show a messagebox when 100% done.
Callback method frameChange():

Put @param tags in correct order (they all got reversed in translation.)
Name parameters in stub definition correctly (to match docstring and
examples.)
(Yardoc generates a warnings for the first two.)
One of the argument names (to_scene) in the docstring was not
code-emphasized.
Revamp method example so it will not choke on the PC.
It causes a MethodError because nil (from_scene) does not respond to
:name.
Add in a if conditional to only print the message once, ...
then thereafter print a nice formatted percentage.